+++ Data/Repositories/Interfaces/Derived/IGamesRepository.cs
@@ -5,5 +5,5 @@ namespace Data.Repositories.Interfaces.Derived;
 
 public interface IGamesRepository : IRepository<Game, AddGameModel, UpdateGameModel>
 {
-    public Task<IEnumerable<Game>> GetByParametersAsync(long[]? genresIds, long[]? platformsIds, long[]? years, long[]? developersIds, long[]? publishersIds, int skip, int take);
+    public Task<IEnumerable<Game>> GetByParametersAsync(long[]? genresIds, long[]? platformsIds, int[]? years, long[]? developersIds, long[]? publishersIds, int skip, int take);
 }
